+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Versions
 * Drop JDK 10 support
 * Migrate Java buildboxes to adoptopenjdk
 * AWS image now uses python3
+* Add new dockerfile "terraform" version 0.12.0
 
 2019-04-05
 ----------