+++ CHANGELOG.md
@@ -19,6 +19,9 @@ Versions
 * Fix Java 8/11 image issues from debian:stretch-slim
 * Drop JDK 10 support
 * Migrate Java buildboxes to adoptopenjdk
+* AWS image now uses python3
+* Add new dockerfile "terraform" version 0.12.0
+* Fix Golang version for accurate image tag
 * Add a new image integrating NCC Scoutsuite
 
 2019-04-05