+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Versions
 * Add new dockerfile "terraform" version 0.12.0
 * Fix Golang version for accurate image tag
 * Add a new image integrating NCC Scoutsuite
+* Change Ansible base image for python:2.7-alpine3.8
 
 2019-04-05
 ----------